part 4 finish

+++ part 4/blogList/controllers/blogs.js
@@ -1,20 +1,59 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const jwt = require('jsonwebtoken')
+const User = require('../models/user')
+const config = require('../utils/config')
+
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
         response.json(blogs)
 
 })
 
 blogsRouter.post('/', async (request, response) => {
-    const blog = new Blog(request.body)
+    const body = request.body
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
+
+    if (!user) {
+        return response.status(400).json({ error: 'UserId missing or not valid' })
+    }
+
+    const blog = new Blog({
+        title: body.title,
+        author: body.author,
+        url: body.url,
+        likes: 0,
+        user: user._id
+    })
 
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
 
     response.status(201).json(savedBlog)
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
+    const user = request.user // From userExtractor
+
+    if (!user) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const blog = await Blog.findById(request.params.id)
+
+    if (!blog) {
+        return response.status(204).end() // Already gone
+    }
+
+        if (blog.user.toString() !== user.id.toString()) {
+        return response.status(401).json({ error: 'only the creator can delete this blog' })
+    }
     await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
 })
